test(ui): fold the paging-anchor regression into the scroll tests
The standalone paging-reading-anchor file repeated the global
save/restore, React root lifecycle, LinkeDOM setup and environment
wiring already present in use-chat-scroll.test.tsx. The case now runs
there under a shared file-local environment installer, keeping the
no-scroll-event edge, the real hook and authority composition, and the
ordering assertion that the anchor is reported before
onLoadEarlierHistory runs (removing the fix still fails it).

Generated-by: Claude Code (Opus 5)
